+++ geoapi/models/__init__.py
@@ -6,5 +6,6 @@ from .users import User
 from .overlay import Overlay
 from .tile_server import TileServer
 from .observable_data import ObservableDataProject
-from .notification import Notification
+from .notification import Notification, ProgressNotification
 from .imported_file import ImportedFile
+from .streetview import Streetview, StreetviewInstance, StreetviewSequence, StreetviewOrganization