+++ geoapi/settings.py
@@ -16,6 +16,7 @@ class Config(object):
     RABBITMQ_HOSTNAME = 'rabbitmq'
     RESTPLUS_MASK_SWAGGER = False
     TENANT = os.environ.get("TENANT")
+    STREETVIEW_DIR = os.environ.get('STREETVIEW_DIR', '/assets/streetview')
 
 class ProductionConfig(Config):
     DEBUG = False
@@ -27,11 +28,13 @@ class ProductionConfig(Config):
     RABBITMQ_PASSWD = os.environ.get("RABBITMQ_PASSWD")
     RABBITMQ_VHOST = os.environ.get("RABBITMQ_VHOST")
     RABBITMQ_HOSTNAME = os.environ.get("RABBITMQ_HOSTNAME")
+    MAPILLARY_CLIENT_ID = os.environ.get('MAPILLARY_CLIENT_ID', 'MLY|5156692464392931|4f1118aa1b06f051a44217cb56bedf79')
 
 
 class DevelopmentConfig(Config):
     DEBUG = True
     PROPAGATE_EXCEPTIONS = False
+    MAPILLARY_CLIENT_ID = os.environ.get('MAPILLARY_CLIENT_ID', 'MLY|4866220476802272|cedfb10deac752ca3ddf83997cef60a4')
 
 class TestingConfig(Config):
     DB_USERNAME = 'dev'