+++ geoapi/tests/api_tests/test_projects_routes.py
@@ -80,7 +80,7 @@ def test_project_permissions(test_client, projects_fixture):
 
 def test_project_protected(test_client, projects_fixture):
     u2 = db_session.query(User).get(2)
-    resp = test_client.get('/projects/1/', headers={'x-jwt-assertion-test': u2.jwt})
+    resp = test_client.get(f'/projects/{projects_fixture.id}/', headers={'x-jwt-assertion-test': u2.jwt})
     assert resp.status_code == 403
 
 
@@ -88,28 +88,36 @@ def test_project_data(test_client, projects_fixture):
     u1 = db_session.query(User).get(1)
     resp = test_client.get('/projects/', headers={'x-jwt-assertion-test': u1.jwt})
     data = resp.get_json()
-    assert data[0]["name"] == "test"
-    assert data[0]["description"] == "description"
+    assert data[0]["name"] == projects_fixture.name
+    assert data[0]["description"] == projects_fixture.description
+
+
+def test_project_data_single(test_client, projects_fixture):
+    u1 = db_session.query(User).get(1)
+    resp = test_client.get(f'/projects/{projects_fixture.id}/', headers={'x-jwt-assertion-test': u1.jwt})
+    data = resp.get_json()
+    assert data["name"] == projects_fixture.name
+    assert data["description"] == projects_fixture.description
 
 
 def test_project_data_protected(test_client, projects_fixture):
     u2 = db_session.query(User).get(2)
-    resp = test_client.get('/projects/{}/'.format(projects_fixture.id), headers={'x-jwt-assertion-test': u2.jwt})
+    resp = test_client.get(f'/projects/{projects_fixture.id}/', headers={'x-jwt-assertion-test': u2.jwt})
     assert resp.status_code == 403
 
 
 def test_project_data_allow_public_access(test_client, public_projects_fixture):
-    resp = test_client.get('/projects/{}/'.format(public_projects_fixture.id))
+    resp = test_client.get(f'/projects/{public_projects_fixture.id}/')
     assert resp.status_code == 200
 
     u2 = db_session.query(User).get(2)
-    resp = test_client.get('/projects/{}/'.format(public_projects_fixture.id), headers={'x-jwt-assertion-test': u2.jwt})
+    resp = test_client.get(f'/projects/{public_projects_fixture.id}/', headers={'x-jwt-assertion-test': u2.jwt})
     assert resp.status_code == 200
 
 
 def test_delete_empty_project(test_client, projects_fixture):
     u1 = db_session.query(User).get(1)
-    resp = test_client.delete('/projects/1/', headers={'x-jwt-assertion-test': u1.jwt})
+    resp = test_client.delete(f'/projects/{projects_fixture.id}/', headers={'x-jwt-assertion-test': u1.jwt})
     assert resp.status_code == 200
     proj = db_session.query(Project).get(1)
     assert proj is None
@@ -117,14 +125,14 @@ def test_delete_empty_project(test_client, projects_fixture):
 
 def test_delete_unauthorized(test_client, projects_fixture):
     u2 = db_session.query(User).get(2)
-    resp = test_client.delete('/projects/1/', headers={'x-jwt-assertion-test': u2.jwt})
+    resp = test_client.delete(f'/projects/{projects_fixture.id}/', headers={'x-jwt-assertion-test': u2.jwt})
     assert resp.status_code == 403
     proj = db_session.query(Project).get(1)
     assert proj is not None
 
 
 def test_delete_unauthorized_guest(test_client, projects_fixture):
-    resp = test_client.delete('/projects/1/')
+    resp = test_client.delete(f'/projects/{projects_fixture.id}/')
     assert resp.status_code == 403
     proj = db_session.query(Project).get(1)
     assert proj is not None
@@ -133,7 +141,7 @@ def test_delete_unauthorized_guest(test_client, projects_fixture):
 def test_add_user(test_client, projects_fixture):
     u1 = db_session.query(User).get(1)
     resp = test_client.post(
-        '/projects/1/users/',
+        f'/projects/{projects_fixture.id}/users/',
         json={"username": "newUser"},
         headers={'x-jwt-assertion-test': u1.jwt}
     )
@@ -143,7 +151,7 @@ def test_add_user(test_client, projects_fixture):
 def test_add_user_unauthorized(test_client, projects_fixture):
     u2 = db_session.query(User).get(2)
     resp = test_client.post(
-        '/projects/1/users/',
+        f'/projects/{projects_fixture.id}/users/',
         json={"username": "newUser"},
         headers={'x-jwt-assertion-test': u2.jwt}
     )
@@ -152,7 +160,7 @@ def test_add_user_unauthorized(test_client, projects_fixture):
 
 def test_add_user_unauthorized_guest(test_client, projects_fixture):
     resp = test_client.post(
-        '/projects/1/users/',
+        f'/projects/{projects_fixture.id}/users/',
         json={"username": "newUser"},
     )
     assert resp.status_code == 403
@@ -161,19 +169,19 @@ def test_add_user_unauthorized_guest(test_client, projects_fixture):
 def test_delete_user(test_client, projects_fixture):
     u1 = db_session.query(User).get(1)
     resp = test_client.post(
-        '/projects/1/users/',
+        f'/projects/{projects_fixture.id}/users/',
         json={"username": "newUser"},
         headers={'x-jwt-assertion-test': u1.jwt}
     )
     assert resp.status_code == 200
-    resp = test_client.delete('/projects/1/users/newUser/',
+    resp = test_client.delete(f'/projects/{projects_fixture.id}/users/newUser/',
                               headers={'x-jwt-assertion-test': u1.jwt})
     assert resp.status_code == 200
 
 
 def test_delete_user_unauthorized(test_client, projects_fixture):
     u2 = db_session.query(User).get(2)
-    resp = test_client.delete('/projects/1/users/test1/',
+    resp = test_client.delete(f'/projects/{projects_fixture.id}/users/test1/',
                               headers={'x-jwt-assertion-test': u2.jwt})
     assert resp.status_code == 403
 
@@ -184,7 +192,7 @@ def test_delete_user_unauthorized(test_client, projects_fixture):
 def test_upload_gpx(test_client, projects_fixture, gpx_file_fixture):
     u1 = db_session.query(User).get(1)
     resp = test_client.post(
-        '/projects/1/features/files/',
+        f'/projects/{projects_fixture.id}/features/files/',
         data={"file": gpx_file_fixture},
         headers={'x-jwt-assertion-test': u1.jwt}
     )
@@ -194,7 +202,7 @@ def test_upload_gpx(test_client, projects_fixture, gpx_file_fixture):
 def test_upload_image(test_client, projects_fixture, image_file_fixture):
     u1 = db_session.query(User).get(1)
     resp = test_client.post(
-        '/projects/1/features/files/',
+        f'/projects/{projects_fixture.id}/features/files/',
         data={"file": image_file_fixture},
         headers={'x-jwt-assertion-test': u1.jwt}
     )
@@ -204,7 +212,7 @@ def test_upload_image(test_client, projects_fixture, image_file_fixture):
 def test_import_image_tapis(test_client, projects_fixture, import_file_from_agave_mock):
     u1 = db_session.query(User).get(1)
     resp = test_client.post(
-        '/projects/1/features/files/import/',
+        f'/projects/{projects_fixture.id}/features/files/import/',
         json={"files": [{"system": "designsafe.storage.default", "path": "file.jpg"}]},
         headers={'x-jwt-assertion-test': u1.jwt}
     )
@@ -213,7 +221,8 @@ def test_import_image_tapis(test_client, projects_fixture, import_file_from_agav
 
 def test_get_point_clouds_listing(test_client, projects_fixture, point_cloud_fixture):
     u1 = db_session.query(User).get(1)
-    resp = test_client.get('/projects/1/point-cloud/', headers={'x-jwt-assertion-test': u1.jwt})
+    resp = test_client.get(f'/projects/{projects_fixture.id}/point-cloud/',
+                           headers={'x-jwt-assertion-test': u1.jwt})
     assert resp.status_code == 200
     data = resp.get_json()
     assert len(data) == 1
@@ -221,13 +230,14 @@ def test_get_point_clouds_listing(test_client, projects_fixture, point_cloud_fix
 
 def test_get_point_cloud(test_client, projects_fixture, point_cloud_fixture):
     u1 = db_session.query(User).get(1)
-    resp = test_client.get('/projects/1/point-cloud/1/', headers={'x-jwt-assertion-test': u1.jwt})
+    resp = test_client.get(f'/projects/{projects_fixture.id}/point-cloud/1/',
+                           headers={'x-jwt-assertion-test': u1.jwt})
     assert resp.status_code == 200
 
 
 def test_get_project_features_empty(test_client, projects_fixture):
     u1 = db_session.query(User).get(1)
-    resp = test_client.get('/projects/1/features/',
+    resp = test_client.get(f'/projects/{projects_fixture.id}/features/',
                            headers={'x-jwt-assertion-test': u1.jwt})
     assert resp.status_code == 200
 
@@ -244,7 +254,7 @@ def test_get_project_features_empty_public_access(test_client, public_projects_f
 
 def test_get_project_features_single_feature(test_client, projects_fixture, feature_fixture):
     u1 = db_session.query(User).get(1)
-    resp = test_client.get('/projects/1/features/',
+    resp = test_client.get(f'/projects/{projects_fixture.id}/features/',
                            headers={'x-jwt-assertion-test': u1.jwt})
     data = resp.get_json()
     assert resp.status_code == 200
@@ -261,7 +271,7 @@ def test_get_project_features_single_feature_public_access(test_client, public_p
 def test_get_project_features_filter_with_assettype(test_client, projects_fixture,
                                                     feature_fixture, image_feature_fixture):
     u1 = db_session.query(User).get(1)
-    resp = test_client.get('/projects/1/features/',
+    resp = test_client.get(f'/projects/{projects_fixture.id}/features/',
                            query_string={'assetType': 'image'},
                            headers={'x-jwt-assertion-test': u1.jwt})
     data = resp.get_json()
@@ -273,7 +283,7 @@ def test_get_project_features_filter_with_bounding_box(test_client, projects_fix
                                                        feature_fixture, image_feature_fixture):
     bbox = [-80.9, 32.61, -80, 32.62]
     u1 = db_session.query(User).get(1)
-    resp = test_client.get('/projects/1/features/',
+    resp = test_client.get(f'/projects/{projects_fixture.id}/features/',
                            query_string='bbox={}'.format(','.join(map(str, bbox))),
                            headers={'x-jwt-assertion-test': u1.jwt})
     data = resp.get_json()
@@ -285,16 +295,17 @@ def test_get_project_features_filter_with_date_range(test_client, projects_fixtu
     u1 = db_session.query(User).get(1)
     start_date = (datetime.datetime.now()-datetime.timedelta(minutes=5)).isoformat()
     end_date = (datetime.datetime.now()+datetime.timedelta(minutes=5)).isoformat()
-    resp = test_client.get('/projects/1/features/',
+    resp = test_client.get(f'/projects/{projects_fixture.id}/features/',
                            query_string={'startDate': start_date,
                                          'endDate': end_date},
                            headers={'x-jwt-assertion-test': u1.jwt})
     data = resp.get_json()
     assert resp.status_code == 200
+    assert projects_fixture.id == feature_fixture.id
     assert len(data['features']) == 1
 
     start_date = (datetime.datetime.now()+datetime.timedelta(minutes=1)).isoformat()
-    resp = test_client.get('/projects/1/features/',
+    resp = test_client.get(f'/projects/{projects_fixture.id}/features/',
                            query_string={'startDate': start_date,
                                          'endDate': end_date},
                            headers={'x-jwt-assertion-test': u1.jwt})
@@ -306,7 +317,7 @@ def test_get_project_features_filter_with_date_range(test_client, projects_fixtu
 def test_import_shapefile_tapis(test_client, projects_fixture, import_file_from_agave_mock):
     u1 = db_session.query(User).get(1)
     resp = test_client.post(
-        '/projects/1/features/files/import/',
+        f'/projects/{projects_fixture.id}/features/files/import/',
         json={"files": [{"system": "designsafe.storage.default", "path": "file.shp"}]},
         headers={'x-jwt-assertion-test': u1.jwt}
     )
@@ -323,7 +334,7 @@ def test_update_project(test_client, projects_fixture):
         },
     }
     resp = test_client.put(
-        '/projects/1/',
+        f'/projects/{projects_fixture.id}/',
         json=data,
         headers={'x-jwt-assertion-test': u1.jwt}
     )
@@ -339,10 +350,10 @@ def test_update_project_unauthorized_guest(test_client, public_projects_fixture)
         'project': {
             'name': "Renamed Project",
             'description': "New Description",
-        },
+        }
     }
     resp = test_client.put(
-        '/projects/1/',
+        f'/projects/{public_projects_fixture.id}/',
         json=data
     )
     assert resp.status_code == 403
@@ -379,7 +390,6 @@ def test_create_observable_project_already_exists(test_client,
 
 
 def test_create_observable_project(test_client,
-                                   projects_fixture,
                                    get_system_users_mock,
                                    import_from_agave_mock,
                                    agave_utils_with_geojson_file_mock):
@@ -396,28 +406,20 @@ def test_create_observable_project(test_client,
         'watch_content': False
     }
 
-    resp = test_client.post(
-        '/projects/',
-        json=data,
-        headers={'x-jwt-assertion-test': u1.jwt}
-    )
+    resp = test_client.post('/projects/',
+                            json=data,
+                            headers={'x-jwt-assertion-test': u1.jwt})
 
     assert resp.status_code == 200
-    proj = db_session.query(Project).get(2)
+    proj = db_session.query(Project).get(1)
     assert proj.name == "Observable name"
 
 
-def test_make_project_observable(test_client,
-                                 projects_fixture,
-                                 get_system_users_mock,
-                                 observable_projects_fixture,
-                                 import_from_agave_mock,
-                                 agave_utils_with_geojson_file_mock):
-    u1 = db_session.query(User).get(1)
+def test_create_observable_project_unauthorized(test_client):
     data = {
         'project': {
-            'name': "Renamed Project",
-            'description': "New Description",
+            'name': 'Observable name',
+            'description': 'Observable description',
             'system_id': 'testSystem',
             'system_path': 'testPath',
             'system_file': 'testFilename',
@@ -426,39 +428,6 @@ def test_make_project_observable(test_client,
         'watch_content': False
     }
 
-    resp = test_client.put(
-        '/projects/1/',
-        json=data,
-        headers={'x-jwt-assertion-test': u1.jwt}
-    )
+    resp = test_client.post('/projects/', json=data)
 
-    assert resp.status_code == 200
-
-
-def test_make_project_observable_already_exists(test_client,
-                                                projects_fixture,
-                                                get_system_users_mock,
-                                                observable_projects_fixture,
-                                                import_from_agave_mock,
-                                                agave_utils_with_geojson_file_mock):
-    u1 = db_session.query(User).get(1)
-    data = {
-        'project': {
-            'name': "Renamed Project",
-            'description': "New Description",
-            'system_id': observable_projects_fixture.system_id,
-            'system_path': observable_projects_fixture.path,
-            'system_file': 'testFilename',
-        },
-        'observable': True,
-        'watch_content': False
-    }
-
-    resp = test_client.put(
-        '/projects/1/',
-        json=data,
-        headers={'x-jwt-assertion-test': u1.jwt}
-    )
-
-    assert resp.status_code == 409
-    assert "Conflict, a project for this storage system/path already exists" in resp.json['message']
+    assert resp.status_code == 403